+++ izumi-reflect/izumi-reflect/src/test/scala/izumi/reflect/test/SharedTagTest.scala
@@ -825,6 +825,42 @@ abstract class SharedTagTest extends AnyWordSpec with XY[String] with TagAsserti
       assertSameStrict(tag.tag, Tag[BIOService[λ[(E, A) => Either[A, E]]]].tag)
     }
 
+    "handles abstract types instead of parameters" in {
+      trait T1 {
+        type F[F0[_], A0] = OptionT[F0, A0]
+        type C[_, _]
+        type G[_]
+        type A
+        type B
+
+        def x: Tag[F[G, Either[A, B]]]
+      }
+
+      val t1: T1 {
+        type G[T] = List[T]
+        type C[A0, B0] = Either[A0, B0]
+        type A = Int
+        type B = Byte
+      } = new T1 {
+        type G[T] = List[T]
+        type C[A0, B0] = Either[A0, B0]
+        type A = Int
+        type B = Byte
+
+        // Inconsistent handling of type aliases by scalac...
+        // No TagK for G, but if G is inside an object or enclosing class
+        // then there is a TagK
+        val g: TagK[G] = TagK[List]
+
+        final val x: Tag[F[G, Either[A, B]]] = {
+          implicit val g0: TagK[G] = g
+          val _ = g0
+          Tag[F[G, C[A, B]]]
+        }
+      }
+
+      assertSameStrict(t1.x.tag, fromRuntime[OptionT[List, Either[Int, Byte]]])
+    }
   }
 
 }
